refactor: Rename settings utils.ts to operations.ts and add barrel exports
- Rename lib/features/settings/utils.ts → operations.ts for clarity
- Add lib/features/settings/index.ts barrel export
- Update imports in API routes to use new path

Improves module naming consistency (operations vs utils) and provides
clean public API surface.
